nextn/cloud_function_vector_search/requirements.txt
google-cloud-aiplatform>=1.30.0
functions-framework>=3.0.0
# Optional: include requests if you need outbound HTTP calls
requests>=2.28.0
